recommendation_service/requirements.txt
fastapi
uvicorn
requests
pandas
scikit-surprise
